Demo_SimGRAG/pyproject.toml
[project]
name = "demo-simgrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "einops>=0.8.0",
    "fastapi>=0.136.1",
    "gdown>=6.0.0",
    "joblib>=1.5.3",
    "neo4j>=6.2.0",
    "openai>=2.33.0",
    "pandas>=2.3.3",
    "pyarrow>=24.0.0",
    "pydantic>=2.13.3",
    "pymilvus>=2.6.12",
    "pyspark>=4.1.1",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=3.0.0",
    "tqdm>=4.67.3",
    "uvicorn>=0.46.0",
]
